+++ src/sll/debug_data.c
@@ -12,9 +12,10 @@ static sll_node_t* _remove_debug_data(sll_node_t* o){
 		o=(o->t==SLL_NODE_TYPE_CHANGE_STACK?o->dt._p:o+1);
 	}
 	switch (o->t){
-		case SLL_NODE_TYPE_CHAR:
 		case SLL_NODE_TYPE_INT:
 		case SLL_NODE_TYPE_FLOAT:
+		case SLL_NODE_TYPE_CHAR:
+		case SLL_NODE_TYPE_COMPLEX:
 		case SLL_NODE_TYPE_STRING:
 		case SLL_NODE_TYPE_IDENTIFIER:
 		case SLL_NODE_TYPE_FIELD:
@@ -84,9 +85,10 @@ static sll_node_t* _remove_type_names(sll_node_t* o){
 		o=(o->t==SLL_NODE_TYPE_CHANGE_STACK?o->dt._p:o+1);
 	}
 	switch (o->t){
-		case SLL_NODE_TYPE_CHAR:
 		case SLL_NODE_TYPE_INT:
 		case SLL_NODE_TYPE_FLOAT:
+		case SLL_NODE_TYPE_CHAR:
+		case SLL_NODE_TYPE_COMPLEX:
 		case SLL_NODE_TYPE_STRING:
 		case SLL_NODE_TYPE_IDENTIFIER:
 		case SLL_NODE_TYPE_FIELD:
@@ -140,7 +142,7 @@ static sll_node_t* _remove_type_names(sll_node_t* o){
 				return o;
 			}
 		case SLL_NODE_TYPE_DECL:
-			o->dt.d.nm=SLL_MAX_STRING_INDEX;
+			o->dt.dc.nm=SLL_MAX_STRING_INDEX;
 			break;
 	}
 	sll_arg_count_t l=o->dt.ac;