+++ src/sll/metadata.c
@@ -23,9 +23,10 @@ static sll_node_t* _mark(sll_node_t* o,bitmap_t* m){
 		}
 	}
 	switch (o->t){
-		case SLL_NODE_TYPE_CHAR:
 		case SLL_NODE_TYPE_INT:
 		case SLL_NODE_TYPE_FLOAT:
+		case SLL_NODE_TYPE_CHAR:
+		case SLL_NODE_TYPE_COMPLEX:
 		case SLL_NODE_TYPE_IDENTIFIER:
 		case SLL_NODE_TYPE_FUNCTION_ID:
 			return o+1;
@@ -62,10 +63,10 @@ static sll_node_t* _mark(sll_node_t* o,bitmap_t* m){
 			}
 		case SLL_NODE_TYPE_DECL:
 			{
-				if (o->dt.d.nm!=SLL_MAX_STRING_INDEX){
-					*(m+(o->dt.d.nm>>6))|=1ull<<(o->dt.d.nm&63);
+				if (o->dt.dc.nm!=SLL_MAX_STRING_INDEX){
+					*(m+(o->dt.dc.nm>>6))|=1ull<<(o->dt.dc.nm&63);
 				}
-				sll_arg_count_t l=o->dt.d.ac;
+				sll_arg_count_t l=o->dt.dc.ac;
 				o++;
 				while (l){
 					l--;
@@ -98,9 +99,10 @@ static sll_node_t* _update(sll_node_t* o,sll_string_index_t* sm){
 		}
 	}
 	switch (o->t){
-		case SLL_NODE_TYPE_CHAR:
 		case SLL_NODE_TYPE_INT:
 		case SLL_NODE_TYPE_FLOAT:
+		case SLL_NODE_TYPE_CHAR:
+		case SLL_NODE_TYPE_COMPLEX:
 		case SLL_NODE_TYPE_IDENTIFIER:
 		case SLL_NODE_TYPE_FUNCTION_ID:
 			return o+1;
@@ -137,10 +139,10 @@ static sll_node_t* _update(sll_node_t* o,sll_string_index_t* sm){
 			}
 		case SLL_NODE_TYPE_DECL:
 			{
-				if (o->dt.d.nm!=SLL_MAX_STRING_INDEX){
-					o->dt.d.nm=*(sm+o->dt.d.nm);
+				if (o->dt.dc.nm!=SLL_MAX_STRING_INDEX){
+					o->dt.dc.nm=*(sm+o->dt.dc.nm);
 				}
-				sll_arg_count_t l=o->dt.d.ac;
+				sll_arg_count_t l=o->dt.dc.ac;
 				o++;
 				while (l){
 					l--;